clazy: Don't include the entire QtCore module in websocket.cpp.
Including the <QtCore> module header file causes the compiler to
include at least 140 other header files, likely more.  Remove this
include and replace it with an include of the header that is actually
needed by the code.

The only reason for this include is to get the qPow function, which is
nothing more than a wrapper around std::pow.  Replace all three calls
to qPow with std::pow and include cmath instead of QtCore.