Ignore capture cards without video source
Ignore the capture cards that are not connected to a video source.
This is for most users a situation that never happens, and there are
messages given by the backend at startup when there are capture cards
without a video source as this might indicate a configuration error.
Howver, test systems can have many capture cards of which only a few
are to be used simultaneously and disconnecting the video source is
the easiest way to temporarily disable a capture card.