EIT active scan continue with next transport at each start
The first time an active EIT scan is started on a capture card
the list with transports and channels to be scanned is created
and a random start channel is selected.
Previously, on each subsequent active scan start again a
random start channel was selected.
Now on each subsequent active scan start scanning continues
at the next channel.
This makes the scanning process better in case scanning is
interrupted, for example by a recording on that capture card
or by doing EIT scanning on conflicting inputs.