EIT active scan on conflicting inputs
EIT active scan is now supported on conflicting inputs.
Conflicting inputs are inputs that have independent video sources
but that share a tuner card, such as a DVB-S/S2 tuner card that
connects to multiple satellite LNBs with a DiSEqC switch.
On a system with such a configuraton the active EIT scan
is done for a fixed amount time, called the EIT Scan Period,
on each input before moving to the next input.
The next input is currently selected at random. It would be better
to have a deterministic ordering but the current implementation
performs well.
There is a new EIT configuration item added for the
EIT Scan Period, with a default value of 15 minutes.