EIT active scan fix for recording start
When stopping EIT active scan on a conflicting input, in order to
start a recording or Live TV, do a complete TuningShutdowns instead
of only a CloseChannel. The difference is that the TuningShutdowns
also closes the stream handler which is not done on a CloseChannel.
When the stream handler of the input that did the EIT active scan
remains active it will receive packets from the conflicting input
that is selected for the recording or the Live TV.
This then caused a few "Program not found in PAT" error messages.