Check network ID for Freesat EIT PID
The UK TV on satellite Astra-2, also known as Freesat, uses the non-standard
PID value 0xF02 (3842) for EIT. Check that the network ID is indeed the network
ID used by Freesat before adding the Freesat EIT PID value to the list of PIDs
to listen to.
This fixes an issue for some channels on the Hotbird 13 satellite who were
received with an additional stream from another channel on the same multiplex
that had the Freesat EID PID value.