EIT active scan dynamically limit scan time per transport
In EIT active scan, move to the next transport when no events
have been received for 60 seconds or when the scan time is up.
This improves the scanning efficiency on satellite channels
where many transport stream mulitplexes do have EIT information
but also many do not.
Note that the scan time per transport can be configured
in page mythtv-setup -> General -> EIT Scanner Options
field "EIT transport timeout (mins)".
The default value for this parameter is 5 minutes; the configured
value is now used as the maximum time spent on one transport
before moving to the next transport.