Avoid unintuitive behaviour when the user selects a forced subtitle track
By default, TOGGLECC toggles between displaying the user's selected track
and the forced track (if present). If the user selects a forced track, that
behaviour can make TOGGLECC look like it does nothing. This commit special
cases that scenario, allowing TOGGLECC to toggle between on and off for
the user selected track.