DB Schema Update: Change lambda capture from 'by value' to 'by reference'
This fix corrects a program abort when updating from db version 1302 to 1303.
The problem is that the 'ORDER BY' column name is not specified, causing a
fatal SQL error. The column name should be either 'starttime' or '-starttime'
depending on your time zone.

This occurs because the lambda capture is 'by value' and takes on the value
at the time it was created (which is empty). Further updates to the local
variable are not available to the lambda expression.

Switching the lambda capture to 'by reference' resolves this issue.

Signed-off-by: Klaas de Waal <klaas@kldo.nl>
(cherry picked from commit 7ec562f39820b92c672d430058d2c81cf4844bbd)
Signed-off-by: Klaas de Waal <klaas@kldo.nl>