EIT cache persistency optional
Make it optional to save the content of the EIT cache in the database.
Saving the content of the EIT cache in the database and using that at
the next start of the backend is the existing behavior.
This reduces EIT event processing at a restart of the backend but only when
there is only a short time period in between. Also, this comes at the cost
of updating the copy of the EIT cache in the database continuously.
When the time period is significant, think of more than a few hours,
the copied EIT cache may be obsolete already as indicated by the table version.
Given that the version field is only 5 bits, the value does wrap around
which means that, if the backend has been offline for a few days, there is
the possibility that invalid cached EIT cache data is seen as valid.
The default behavior is to have the persistency enabled because that
is how it has been since 2006.