EIT event processing rate configurable
The rate at which EIT events are processed in the backend is limited so
that there is always enough processing capacity for the other backend tasks.
For many years this limit is set to 20 events per 400 milliseconds.
Measuring CPU load on recent hardware shows that mythbackend can process
events at a much higher rate without a noticeable increase of CPU load.
A higher event procesing rate makes it possible to update the EPG faster.
The default value remains at 20 events but the value can now be
configured from 20 to 1000 events per 400 milliseconds.