metadata/Music: remove unused external dependency on simplejson
The simplejson module was added to the python standard library in 2.6;
here, it is inconsistently imported for python 2.6 (where it already
exists) and lower (where it does not) instead of using `json`.

But, mythtv only supports python 3.8 for a while now. So this
conditional is merely dead code. The import name is still aliased using
`as simplejson`, though, which is bad for visibility and analysis.