Extend message area in fullscreen MythDialogBox in MythCenter-wide
Sometimes a program cannot be recorded because there are not
enough tuners available and other programs have higher recording
priorities. The other programs that will be recorded instead are shown
in the Upcoming Recordings by pressing Enter on the conflicting recording.
In theme MythCenter-wide the message area had only room to show
two conflicting recordings. The message area is now extended
to show a few more conflicting recordings.
In the Terra theme the MythDialogBox is configured so that there is
already enough room to show many conflicting recordings.