Python Bindings: Enforce python pip for python 3.12
In case of python 3.12+, only use python-pip to install
the Python Bindings (instead of using `setup.py` directly).

Otherwise, one may get a mixture of the installs of the
Python Bindings done by `setup.py` from the `setuptools`,
which produces an `python egg` installable, and the Bindings
installed by `pip install` which uses a `python wheel`
as install medium.

Fixes #848