Display of OpenGL RGBA parameters wrong
In the OpenGL debug output the number of bits in each pixel for R, G and B is given.
Instead of the bits for B (blue) the number of bits for G (green) is repeated.
This is now fixed by replacing the 2nd greenBufferSize with blueBufferSize.
Thanks to Martin Bene.