Support displaying bookmark state and watched progress in video manager (#742)
* Add support for MARK_TOTAL_FRAMES for videos

Extend ProgramInfo to load, delete and insert
MARK_TOTAL_FRAMES entries in filemarkup table.
This markup type can now be written by mythcommflag
when rebuilding the seektable for a video file.
This is in preparation to support displaying watched
progress within the video manager.

* Save total frame count for video files

When exiting the media player, check if the current
video file alreadey has a total frame count saved
within the filemarkup table. If it is not yet present,
then save the total frame count from the player.
This way the watched progress does not need extra
tooling to provide the total frame count of a video file.

* Pull playback state for all videos from database

Query and cache playback state related data from
database so the watched progress can be shown for
multiple video list items.

* Add bookmark and watched percent support to videos

Add support for displaying a bookmark state within
the video manager. This uses the playback state class
for caching the playback state for all videos.
Also adds support for displaying watched progress
state to video manager. This needs the last playing
position (already supported) as well as a total
frames markup datum within the filemarkup table.
The watched progress is then calculated and can be
displayed via a progress in the currently selected list
item.

* Signal changed playback state

Signal removal of last played position and bookmark
to video dialog, so it can update the respective
button list item.

* Fix displaying watched state after playback

When watching a video to the end, the player can mark
(if enabled by a setting) the video as watched. This
was not reflected in the GUI state until now.
So after receiving the playback stopped event, we load
the video metadata of the current video and sync it with
the metadata from the cache.

* Add setting to always show watched progress

Adds a setting to always show watched progress even
if the recording or video is marked as watched.
Make behaviour of displaying the watched progress
the same in recordings and video lists.