Add "Audio Language" selection
Add a selection for the preferred audio language.
The existing "Language" setting is used for the menu language
but was until this fix also used as preferred audio language.
When there is no "Audio Language" preference specified
then the "Language" setting is used so users do not need
to configure anything to keep the current behavior.
Note that TVs sold in Europe generally have separate
settings for "Menu Language" and "Audio Language".