Fix segfault in EIT with empty video source
The backend segfaults when EIT is enabled and it is started
with a video source without channels.
Fixed by adding an additional test on the number of channels present.
Note that it can be considered a bug that the EIT scan is started
when there is not a single channel present.

(cherry picked from commit 36d76494c1222311d49c9fa84312c69ef2526c45)