tidy: Fixes for "assignment in if statement" warnings. (plugins)
These assignments are bug-prone because they may have been intended as
equality tests.  Moving the assignments outside of the if/then/else
logic makes the decision path becomes easier to follow.

One fix in this commit is simply to pull the assignment out of the
conditional statement, but the code is in the middle of a large macro.
The other commit rewrites a short if/elseif chain to make the code
easier to understand.

https://clang.llvm.org/extra/clang-tidy/checks/bugprone/assignment-in-if-condition.html