tidy: Fixes for "assignment in if statement" warnings. (programinfo.cpp)
These assignments are bug-prone because they may have been intended as
equality tests.  Moving the assignments outside of the if/then/else
logic makes the decision path becomes easier to follow.

Rewrite the ProgramInfo::QueryStartMark function to convert a chain of
"if/elseif" statements into a simpler chain of "if" statements.  This
removes all the assignments from the middle of the conditional
statements. This should make the code easier to understand.

https://clang.llvm.org/extra/clang-tidy/checks/bugprone/assignment-in-if-condition.html