tidy: Fixes for "assignment in if statement" warnings. (musicmetadata.cpp)
These assignments are bug-prone because they may have been intended as
equality tests.  Moving the assignments outside of the if/then/else
logic makes the decision path becomes easier to follow.

Rewrite the MusicMetadata::getAlbumArtFile function from a chain of
"if/elseif" statements with embedded assignment statements, into a
much simpler loop.  This should make the code easier to understand.

https://clang.llvm.org/extra/clang-tidy/checks/bugprone/assignment-in-if-condition.html