tidy: Fixes for "assignment in if statement" warnings. (libs 2)
These assignments are bug-prone because they may have been intended as
equality tests.  Moving the assignments outside of the if/then/else
logic makes the decision path becomes easier to follow.

These fixes involve simple rewrites of "if" statements to pull an
assignment statement out of the middle of a complex conditional.  This
should make the code easier to understand.  They do move a call to
QString::indexOf before a conditional so that it will be executed when
it might not have been called before.  This should be harmless, and
doing it this way reduces the size of the necessary change.

https://clang.llvm.org/extra/clang-tidy/checks/bugprone/assignment-in-if-condition.html