Change ATSC major/minor channel separator from "_" to "."
With HDHomeRun tuners it is possible to import the channel list directly
via an XML or an M3U file. In these files the separator between the
major and minor channel number is the dot "." and this gives then
channel numbers as "2.1" or "23.11".
Channel scanning on ATSC streams with DVB tuners and also with HDHomeRun
tuners using the libhdhomerun transport stream Interface, gives the channel
numbers as separate major and minor integer numbers. Currently MythTV adds
the underscore character "_" as separator between the major and minor channel
number, e.g. "2_1" or "23_11", in creating the channel number strings
that are stored in the database.
It is an undesirable situation that channels from a HDHomeRun do have
different channel numbers depending on the way the channels are retrieved.
Also, it has been mentioned that the dot is preferred as separator because
Schedules Direct does also use the dot, so matching a Schedules Direct lineup
with the channels found by MythTV channel scanning is easier when using dots.
Altogether enough reasons to change the separator from "_" to ".".

Refs #864