tidy: Fixes for "assignment in if statement" warnings. (libs 3)
These assignments are bug-prone because they may have been intended as
equality tests.  Moving the assignments outside of the if/then/else
logic makes the decision path becomes easier to follow.

Rewrites an "if" statements to pull two assignment statement out of
the middle of a complex conditional.  This should make the code easier
to understand.

https://clang.llvm.org/extra/clang-tidy/checks/bugprone/assignment-in-if-condition.html