tidy: Convert #defines to static constexpr.
The tidy macro-to-enum checker identified these macros as being
eligible to be converted to an enum, but they aren't related.  A
better solution is to convert them to static constexpr.  This means
that they are no longer blind text substitutions by the preprocessor,
but now have a type and are processed by the c++ compiler.