tidy: Replace "length() == 0" checks with "empty()".
The emptiness of a container should be checked using the empty()
method instead of the size()/length() method. It is not guaranteed
that size()/length() is a constant-time function, and it is generally
more efficient and also shows clearer intent to use empty().

These changes were made by the clang-tidy program.

https://clang.llvm.org/extra/clang-tidy/checks/readability/container-size-empty.html