Fix compiler warning about disallowed template-id.
The gcc-14 compiler is issuing the warning that "template-id not
allowed for constructor in C++20".

Starting with C++17, the compiler automatically deduces the required
type for constructor, but it was OK for the code to also specify the
type.  In C++20, the code also specifying the type will apparently
become an error.  This commit removes the redundant type
specification.