Use the right variable for choosing system vs local exiv2.
The configure file uses the label using_system_libexiv2, while a
number of the test.pro files use some variation on this name.  This
never mattered before because everything had to use the local exiv2.
Now that some builds are using the local exiv2 and others are using
the system exiv2, this needs to be corrected.