Fix compilation on FreeBSD 15.
The regular expression to strip the /usr/local/include path from the
FREETYPE_CFLAGS has always been broken, but the break was never
detected because that path always appeared in the middle of the cflags
string.  In FreeBSD 15, that include path now appears at the end of
the cflags string and the broken regex fails to remove it.  Fix the
regular expression to correctly remove the include path at the end of
the cflags string.  This has been tested on FreeBSD 13, 14, and 15.