tidy: Fix warning about bit shifting in websocket.cpp.
The clang-tidy core.BitwiseShift pointed out undefined behavior
processing frames in the websocket code.  The payload header
processing code is stating that a value is 8-bits wide and then
shifting it by anywhere from 0 to 56 bits.  This is technically
undefined but probably works for shifts up to 24 bits (because the
8-bit value is held in a 32-bit register), but probably fails for
larger shifts.  Forcing the compiler to expand the 8-bit value into a
64-bit register before shifting (instead of the current implicit
expansion after shifting) prevents the warning messages.

https://clang.llvm.org/docs/analyzer/checkers.html#core-bitwiseshift