tidy: Simple fixes for "assignment in if statement" warnings. (netvision)
These assignments are bug-prone because they may have been intended as
equality tests.  Moving the assignments outside of the if/then/else
logic makes the decision path becomes easier to follow.

This commit contains only fixes where it is simple to pull the
assignment out of the conditional statement.  They should make the
code easier to understand and less likely to hide bugs.

https://clang.llvm.org/extra/clang-tidy/checks/bugprone/assignment-in-if-condition.html