tidy: Revert a fix for redundant member initialization.
The change to this macro to remove the redundant initialization caused
a larger set of warnings about missing member initializers.  Restore
the member initialization and ignore the redundant initialization
messages.  A better solution would be to have two different macros,
one for simple types and one for complex types, but that would require
a much larger change.

This reverts commit 3888516c1b2b60357ec1b81fa1cae78219984d70.