Handle HTTP parameter type std::chrono::seconds
The timeout/duration value in MythFrontendService::SendNotification
is expressed as a number of seconds. This number used to be passed as
an integer but this has been changed to std::chrono::seconds in 2021.
In the processing of HTTP SendNotification messages the
std::chrono::seconds parameter type is not recognized, leading to
an uninitialized value used as timeout value.
This is fixed by checking for QMetaType "std::chrono::string" and
copying the parameter accordingly.
There is now also a warning given if a QMetaType is passed
that is not recognized and that cannot be handled.