Use FFmpeg MPEG-TS demuxer when MythTV demuxer fails
The FFmpeg demuxer may fail when there is a change in the streams halfway playback.This can happen
when changing to a commercial and back or when a recording extends through multiple programs
The MythTV demuxer can handle these changes. However, some streams are not correctly handled by the MythTV demuxer.
The code is written such that there is an automatic fallback to the original FFmpeg MPEG-TS demuxer when
the MythTV demuxer fails.
In the refactoring of the code for MythTV release v31, in commit 76bca2a, the fallback code is still present
but is, due to a minor coding mistake, never executed. This is now fixed.
Tested as follows:
BBC Three and BBC Four in the UK are off air until 7:00pm when they start broadcasting.
Recording start at 60 seconds before the start of the first program.
This then creates a recording that cannot be demuxed by the MythTV demuxer but which
can be played back correct whne the FFmpeg demuxer is used.