Support USB HID media control keys now common on remotes and keyboards.
While here, also map emacs-like Ctrl+B and Ctrl+F for rewind and fast
foward as on older MCE remotes like ARC-1100.  This required moving
"jump to start" from Ctrl+B to Ctrl+A.  Menu, Info and Delete are also
mapped to the control keys of the colored ARC-1100 buttons.