libavformat/mpegts: rename avpriv_mpegts_parse_*()
This reduces the changes relative to upstream FFmpeg in mpegts.*
and makes it clearer where in FFmpeg our version is used.

The impetus for this change is that the GitHub Actions Fedora CMake
builder is now building libavdevice/iec61883.c which includes and
uses FFmpeg’s mpegts.h, causing it to not compile.