libmyth/audio/freesurround*: silence -Wshadow
freesurround_decoder.cpp: This was identical to a code block before the if.

freesurround.cpp:

The fsurround_params constructor is never called explicitly, so just set
the default values in the struct definition.  Those values are immediately
overwritten in the FreeSurround constructor anyways.

In FreeSurround::putFrames(), copy the shadowed variables into the places
they are used.