Scheduler::UpdateManuals(): prevent scheduling recordings before the given start date
and rewrite the confusing code that calculated the start times.

I thought about adding something to prevent scheduling recordings in the past,
but this function doesn’t have enough information to determine that.

`
if (MythDate::current().toLocalTime().time() > starttime)
{
    offset++;
}
`
The above code would prevent scheduling recordings that should have already
started.  However, the end time may still be in the future, so it should be
scheduled in that case.  Unfortunately, the startdate, starttime, enddate,
and endtime values do not account for starting/ending the recording early/late
(8 hours in either direction from the start and end), so this cannot be
determined from the information the function has.

Therefore, this function will continue to schedule recordings that may be
entirely in the past.