BitReader: add constant k_cacheSizeMax, replace CHAR_BIT
BitReader does not use `char`s, so CHAR_BIT is not the correct
value to use, even though POSIX requires it to be 8.

BitReader uses the fixed width uint8_t type to read memory.
However, the code would probably work with other sizes for
the smallest addressable unit of memory.