BitReader::refill_cache(): remove masking
In any cached BitReader implementation, the behavior is undefined if the data
being read is changed after the BitReader is constructed since the cache may
contain the old data, which could result in the returned read values being a mix
of the old and new data.  To prevent this you would need to invalidate the
cache, which would turn the cached BitReader into an uncached BitReader.
Therefore, a cached BitReader implementation can assume that the data being read
is never modified during its lifetime (which makes any cached data always valid).

Assuming that the value pointed to by m_buffer has not changed since it was last
read into m_cache, the masking is unnecessary since the cleared values are
either equal to those already in m_cache or will be shifted out.

e.g.:
*m_buffer = 0b1101'1011
m_cache = 0b1001011001101010110000000000000000000000000000000000000000000000
m_cacheSize = 19
m_bitIndex = 3

Once shifted *m_buffer becomes
          0b0000000000000000110110110000000000000000000000000000000000000000

The three bits that were formerly cleared are identical to the corresponding
bits in m_cache.