Use close_range() to close large numbers of files when available.
close_range() is a newish, Linux and FreeBSD system call to close a
large range of file descriptors as is commonly done by the child
process after a fork.  Linux now supports extremely large limits and
the traditional, brute force way of closing file descriptors can incur
a several minute delay in those situations.  This change checks for
and uses close_range() when it is available.