Fix MPTS channel creation
MPTS channel creation can be selected at the channel scan; if selected
then one additional channel is created per transport stream which can
be used to make recordings of the complete transport stream.
An MPTS channel is created as a copy of an existing channel in the
same transport stream. It did however happen that during the
post processing the MPTS channel was discarded because of the
"Logical Channel Numbers Required" or the "Complete Scan Data Required"
options when the existing channel did not meet these criteria.
This is now fixed by selecting an existing channel that has
"Complete Scan Data", i.e. it does appear in PAT and PMT, and
by setting a value for the logical channel number.