mpegts-mythtv.c: remove fallback code from mpegts_read_header()
Disabling CRC is from:
https://github.com/MythTV/mythtv/commit/084e3f83557cdd4ba8f86d31173332a1129b1f42
referencing
https://code.mythtv.org/trac/ticket/328

However, the sample does not trigger that fallback.

Deal with incomplete PMT streams in BBC iPlayer IPTV
https://github.com/MythTV/mythtv/commit/c11ee69c81198dc221c874e8132f1f30a385fa63
references
https://code.mythtv.org/trac/ticket/9926

The sample is unavailable and the issue was not in FFmpeg's unmodified demuxer,
so, following subsequent harmonization, I assume it is no longer an issue.  If
it still is, MythTV's other modifications should be looked at first.