tidy: Quiet "implicit pointer conversions of different levels" warnings. (2)
Tidy is pointing out places where a pointer** is being converted to a
pointer*.  Most of these are calls to av_freep() where the memory is
freed and the pointer to it is set to nullptr.  These all appear to be
correct.  The other calls are also to memory freeing functions and are
probably correct as well.

This seems like a reasonable check to keep running, so add explicit
type conversions to the existing calls to the various free functions.

https://clang.llvm.org/extra/clang-tidy/checks/bugprone/multi-level-implicit-pointer-conversion.html