tidy: Quiet enum base size warning for RecStatus.
When building its data, the Qt meta-object compiler unconditionally
casts enum values to a 'uint' using a C-style cast.  If base type of
the enum is smaller than the size of an 'int', this cast is broken for
negative numbers and changes their value.

The cast apparently first changes the representation of a particular
bit value to unsigned, and then expands it to the new bit size by left
padding with zero bits.  (E.G. The decimal number -3 in a int8_t is
binary 11111101.  This gets left padded with zeros to become
00000000'0000000'00000000'11111101.  If this value is later cast back
to an integer, it now has the value 253.)  If the variable size had
been changed first by sign extension before the representation changed
to unsigned, the moc code would have worked.  (Sign extending 11111101
to 32-bits wide gives 11111111'11111111'11111111'11111101.  If this
value is later cast back to an integer, it would still has the value
of -3.)

All of the recent enum base size changes that use an integer type less
that 32 bits were checked to see if they suffer from this problem.
The RecStatus enum is the only one that is registered as a qt
metatype, and is thus the only one affected by this problem.