tidy: Fix warning about reading undefined value from errno.
Clang-tidy is pointing out that calls to unlink() can change the value
of errno, and the test of errno at the end of the function is
potentially testing an undefined value.  Since the code is interested
in the value of errno after the previous open call (unlink doesn't
return EISDIR), save a copy of errno before calling unlink and check
that copy later.

https://clang.llvm.org/docs/analyzer/checkers.html#unix-errno