tidy: Fix two warnings about file stream related errors.
1) The clang-analyzer-unix.Stream checker warns that the file position
of the stream might be "indeterminate" after a failed operation, which
can cause undefined behavior.  The code is correct, but the GCC
attribute indicating that a function never returns isn't being used.
Restoring these markings eliminates the warning.

2) Remove the check for GCC version less than 3 (aka the year 2001).

3) After fixing #1, tidy now indicates there is a potential resource
leak in the error handling.  Add a call to close a file descriptor
before returning.

https://clang.llvm.org/docs/analyzer/checkers.html#unix-stdclibraryfunctions