tidy: Replace some if/then clauses with calls to std::min/max. (programs)
Replace some conditional if/then statements with the equivalent call
to std::max.  This can help clarify reading the code.  If a variable
is tested for both upper and lower bounds, replace both tests with a
call to std::clamp.

std::min and std::max changes made by clang-tidy.  Conversions of
min/max pairs to std::clamp by hand.  The single static_cast is to
support compiling on both qt5 and qt6.

https://clang.llvm.org/extra/clang-tidy/checks/readability/use-std-min-max.html