move pmt_section to AVProgram
This is a MythTV addition that exports the PMT from an MPEG-TS.  However, each
program has its own PMT so having a single pmt_section in AVFormatContext is not
correct when there are multiple programs in the stream.  Previously, each PMT
overwrote the last one.

Adapt AvFormatDecoder to look in AVProgram.

In my limited testing, the PMT did not seem necessary for ATSC captions.

MythAVBufferRef: Satisfy the rule of five; however, since nullptr is now
allowed, the newly defined constructors and operators are unused.  The compiler
generated ones were causing use after free segmentation faults.