tidy: Fix a couple of "narrowing conversions" warning.
Tidy is warning about a couple of uints that are being assigned to
int.  This appears to be a result of an enum size being changed from
uint->uint8_t, and a function that takes that enum as an argument not
also having its size changed.

These were introduced in bbe421d008.