cmake: Update tweaks to fix fedora glslang issues.
Fedora 39 fixed the problem where it couldn't parse the glslang.pc
file at all because of leading spaces, but it introduced a new problem
because it still requires linking against libraries that are no longer
part of the distribution.  Change the first fix to only apply to
Fedora 38 or earlier, and add a fix for the new problem.

View this diff using the "-w" flag.