cmake: Ensure libz.so is used instead of libz.a.
The Android ndk provides libz.a in an architecture specific
subdirectory (e.g. lib/arm-linux-androideabi/libz.a), and libz.so in a
different architecture specific version specific subdirectory
(e.g. lib/arm-linux-androideabi/29/libz.so).  CMake knows how to
handle the former but not the latter.  Copy both of these (and the
include file) into our install directory so that CMake can easily find
and figure out which to use.