AvFormatDecoder::DecodeCCx08(): remove dead code
The code in libavcodec/mpeg12dec.c will never create a cc_type of 2 for SCTE-20
and cc_type is guaranteed to be less than 2 otherwise by the if condition.

AvFormatDecoder::m_invertScteField is now always 0, so remove it.

AvFormatDecoder::m_lastScteField is now set but unused, so remove it.