Fix HDHomeRun Channel Import type of scan
Using the 'HDHomeRun Channel Import' style of
channel scan on an HdHomerun Prime, with a cablecard
installed, ends up adding a single dtv_multiplex
row to the database which gets pointed to by every
channel row added. This caused problems with
the Scheduler later on because channels meant
to be tuned in using the virtual channel number
are not supposed to have any mplexid specified.

This update adds code to check if a cablecard
is installed. If so, the mplexid field of the
created 'channel' row is set to NULL. In
addition, it's necessary to store the virtual
channel number in the 'freqid' field of the
created 'channel' row because that is what's
used to do virtual channel tuning. See ...

HDHRChannel::Tune(const QString &freqid, int)

With these modifications, it is now possible
to tune into channels populated by
'HDHomeRun Channel Import'.

Refs #964