Fix  HDHomeRun Channel Import for Prime with cablecard (#968)
* Updated scheduler to factor reclimit into mplexid_ok

The configuration instructions tell us that multiple
recordings on the same multiplex will only be allowed
if the Max Recordings value is greater than one.
However, the scheduler, decided to use multiplex
recording without considering Max Recordings. This
update factors in the value of Max Recordings.

A couple of log messages have also been updated.
The first used to tell us "comparing with" but
only identified one title. To know what's being
compared, we need both titles. This has been
implemented. Also, the channels for these two
items are now written to the log message. This
may help when trying to see if the channels
share a multiplex.

The second log message suffered from collision
of two strings. "groupmplexid's:" has been
updated to "group, mplexid's:".

Resolves: #964

* Fix HDHomeRun Channel Import type of scan

Using the 'HDHomeRun Channel Import' style of
channel scan on an HdHomerun Prime, with a cablecard
installed, ends up adding a single dtv_multiplex
row to the database which gets pointed to by every
channel row added. This caused problems with
the Scheduler later on because channels meant
to be tuned in using the virtual channel number
are not supposed to have any mplexid specified.

This update adds code to check if a cablecard
is installed. If so, the mplexid field of the
created 'channel' row is set to NULL. In
addition, it's necessary to store the virtual
channel number in the 'freqid' field of the
created 'channel' row because that is what's
used to do virtual channel tuning. See ...

HDHRChannel::Tune(const QString &freqid, int)

With these modifications, it is now possible
to tune into channels populated by
'HDHomeRun Channel Import'.

Refs #964

* Backing out reclimit changes from earlier commit

In commit 4b31dd1, changes to scheduler.h and
scheduler.cpp were added to provide access to
the max recording limit in the scheduler. This
is no longer needed, so those changes are
being backed out.

Refs: #964

* Backing out scheduler log message changes

Log message changes to scheduler.cpp are being
backed out. They will be implemented in another
branch so that they can be submitted under a
separate pull request than the fix for HDHomeRun
Channel Import.

Refs: #964