merge DatabaseParamsCache into MythDB
GetMythDB() will always return a valid pointer to the global MythDB instance.

Some of the functions previously did not update the DatabaseParams in MythDB;
however, this shouldn't make a difference since only libmyth/mythcontext.cpp
ever called MythDB::SetDatabaseParams().

gContext is now unused in the libraries and can be removed.