Harmonize Title-Parsing of Movies and TV-Series
This change respects parentheses, i.e.: curved marks "()"
in the parsed filenames of a movie, thus parsing of the
release year in parentheses will work.
It enables MythTV's movie grabber to extract the release year
from the parsed filename and do a search for title with year.
The correct syntax for a filename is from now on
"A Movie Title (YYYY).mpg" instead of "A Movie Title YYYY.mpg"
as stated in the wiki.
It resolves ambiguous parsing of different movie titles like
`Wonder Woman 1984` and `Wonder Woman`.

Note: For tv-series, it is already allowed to add the year
in the title, like `The Forsyte Saga (1967)`
e.g.: `The Forsyte Saga (1967) 01x01 A Family Festival.mkv`
is correctly parsed.

Fixes MythTV/mythtv#974