tidy: Don't explicitly re-declare default initializers.
If an field is initialized to the default value for a variable type,
you should use the default initializer instead of explicitly declaring
it.  Removing all default initializers makes it obvious where there
are non-default values.  This also reduces duplicate code when there
are multiple constructors for an object, as the variable's default
value is specified only once.

https://clang.llvm.org/extra/clang-tidy/checks/modernize/use-default-member-init.html